+++ rust-toolchain
@@ -1,1 +1,1 @@
-nightly-2021-09-15
+nightly-2022-10-24
