+++ rust_src/.gitignore
@@ -1,5 +1,3 @@
-target
-
 crates/emacs/src/bindings.rs
 crates/emacs/src/definitions.rs
 crates/emacs/src/globals.rs
