--- rust_src/Cargo.toml.in
@@ -1,81 +1,0 @@
-[package]
-authors = ["emacs-ng members"]
-name = "emacsng"
-description = "Experimental emacs fork"
-license = "GPL-3.0"
-version = "0.1.0"
-build = "build.rs"
-edition = "2018"
-
-[workspace]
-members = ["remacs-lib", "crates/*"]
-exclude = ["ng-bindgen"]
-
-[dependencies]
-remacs-lib = { version = "0.1.0", path = "remacs-lib" }
-emacs = { version = "0.1.0", path = "crates/emacs" }
-lisp-util = { version = "0.1.0", path = "crates/lisp_util" }
-lisp-macros = { version = "0.1.0", path = "crates/lisp_macros" }
-ng_async = { version = "0.1.0", path = "crates/ng_async" }
-lsp_json = { version = "0.1.0", path = "crates/lsp_json" }
-git = { version = "0.1.0", path = "crates/git", optional = true }
-ng_module = { version = "0.1.0", path = "crates/ng_module", optional = true }
-js = { version = "0.1.0", path = "crates/js", optional = true }
-webrender = { version = "0.1.0", path = "crates/webrender", optional = true }
-clippy = { version = "*", optional = true }
-
-[build-dependencies]
-clippy = { version = "*", optional = true }
-lazy_static = "1.2"
-libc = "0.2.95"
-regex = "1.1"
-
-[dev-dependencies]
-tempfile = "3.0"
-
-[lib]
-crate-type = ["staticlib"]
-path = "lib.rs"
-
-# Allow unwinding in functions called from C. See
-# https://www.reddit.com/r/rust/comments/565q1f/wrote_a_blogpost_from_my_experiences_the_arduous/d8h053m/
-[profile.dev]
-panic = "abort"
-
-[profile.release]
-panic = "abort"
-codegen-units = 1
-lto = true
-# This reduces the binary size from 125M to 70M due
-# to Deno inclusion
-opt-level = 'z' # Optimize for size
-
-
-[features]
-default = [
-  @CARGO_DEFAULT_FEATURES@
-  @CARGO_CHECKING_FEATURES@
-]
-
-# Compile with C xml2 library support.
-use-xml2 = []
-# Use a window system
-window-system = ["emacs/window-system"]
-# Use the x11 window system
-window-system-x11 = []
-# Use the nextstep window system
-window-system-nextstep = []
-# Use the w32 window system
-window-system-w32 = []
-# Build with git2rs support
-libgit = ["git"]
-# Use the webrender window system
-window-system-webrender = [@WEBRENDER_DEFAULT_FEATURES@]
-# Treat warnings as a build error on Travis.
-strict = []
-# Use JavaScript and Deno
-javascript = ["js"]
-# Build with dynamic modules support's extensions.
-ng-module = ["ng_module"]
-# Enable glyphs debugging code.
-glyph-debug = []
