+++ rust_src/crates/emacs/.gitignore
@@ -1,0 +1,4 @@
+src/definitions.rs
+src/globals.rs
+src/bindings.rs
+build.rs
