+++ rust_src/crates/emacs/Cargo.toml
@@ -1,7 +1,8 @@
 [package]
 name = "emacs"
 version = "0.1.0"
-edition = "2018"
+edition = "2021"
+rust-version = "1.66"
 
 [lib]
 path = "src/lib.rs"
@@ -10,6 +11,11 @@ path = "src/lib.rs"
 libc = "0.2.95"
 lazy_static = "1.4.0"
 memoffset = "0.6.4"
+
+[build-dependencies]
+bindgen = "0.46"
+regex = "1.6"
+libc = "0.2"
 
 [features]
 window-system = []
