+++ rust_src/crates/git/Cargo.toml
@@ -15,3 +15,6 @@ openssl-sys = { version = "0.9" }
 git2 = "0.13"
 libc = "0.2.95"
 lazy_static = "1.2"
+
+[build-dependencies]
+ng-bindgen = { path = "../../ng-bindgen" }
