+++ rust_src/crates/js/Cargo.toml
@@ -21,3 +21,6 @@ serde_json = { version = "1.0", features = ["preserve_order"] }
 rusty_v8 = "0.22.1"
 tokio = { version = "1.4.0", features = ["full"] }
 tokio-rustls = "0.22.0"
+
+[build-dependencies]
+ng-bindgen = { path = "../../ng-bindgen" }
