+++ rust_src/crates/webrender/.gitignore
@@ -1,1 +1,2 @@
 out/
+Cargo.toml
