+++ rust_src/crates/webrender/Cargo.in
@@ -1,0 +1,44 @@
+[package]
+name = "webrender"
+version = "0.1.0"
+edition = "2021"
+build = "build.rs"
+
+[lib]
+path = "src/lib.rs"
+
+[dependencies]
+emacs = { path = "../emacs", features = ["window-system-webrender"] }
+lisp-macros = { path = "../lisp_macros" }
+lisp-util = {  path = "../lisp_util" }
+colors = {  path = "../../../etc/colors" }
+env_logger = "0.9.3"
+log = "0.4.17"
+libc = "0.2.95"
+lazy_static = "1.2"
+webrender = { git = "https://github.com/servo/webrender.git", rev = "ecbe93f7df289db2823c3c1fb748af5f3e4827ea" }
+font-kit = "0.10.0"
+glutin = "0.26"
+gleam = "0.13"
+copypasta = "0.7.1"
+image = "0.23.14"
+app_units = "0.7.1"
+bit-vec = "0.6.3"
+fontdb = "0.5.4"
+ttf-parser = "0.12.3"
+once_cell = "1.8.0"
+tokio = { version = "1.10.0", features = ["rt-multi-thread", "sync", "net", "macros", "time"] }
+futures = "0.3.16"
+
+[build-dependencies]
+ng-bindgen = { path = "../../ng-bindgen" }
+
+[target.'cfg(target_os = "macos")'.dependencies]
+core-foundation = "0.9.2"
+
+[features]
+default = ["wayland", @WEBRENDER_DEFAULT_FEATURES@]
+x11 = ["copypasta/x11", "surfman/sm-x11", "fontdb/fontconfig"]
+wayland = ["copypasta/wayland", "fontdb/fontconfig"]
+capture=["webrender/capture", "webrender/serialize_program"]
+sw_compositor=["webrender/sw_compositor"]
