--- rust_src/crates/webrender/Cargo.toml
@@ -1,37 +1,0 @@
-[package]
-name = "webrender"
-version = "0.1.0"
-edition = "2018"
-build = "build.rs"
-
-[lib]
-path = "src/lib.rs"
-
-[dependencies]
-emacs = { version = "0.1.0", path = "../emacs", features = ["window-system-webrender"] }
-lisp-macros = { version = "0.1.0", path = "../lisp_macros" }
-lisp-util = { version = "0.1.0", path = "../lisp_util" }
-libc = "0.2.95"
-lazy_static = "1.2"
-webrender = { git = "https://github.com/servo/webrender.git", rev = "ecbe93f7df289db2823c3c1fb748af5f3e4827ea" }
-font-kit = "0.10.0"
-glutin = "0.26"
-gleam = "0.13"
-copypasta = "0.7.1"
-image = "0.23.14"
-app_units = "0.7.1"
-bit-vec = "0.6.3"
-fontdb = "0.5.4"
-ttf-parser = "0.12.3"
-once_cell = "1.8.0"
-tokio = { version = "1.10.0", features = ["rt-multi-thread", "sync", "net", "macros", "time"] }
-futures = "0.3.16"
-
-[build-dependencies]
-cargo_toml = "0.10.1"
-
-[features]
-default = ["wayland", "x11"]
-x11 = ["copypasta/x11", "glutin/x11"]
-wayland = ["copypasta/wayland", "glutin/wayland"]
-capture=["webrender/capture", "webrender/serialize_program", "webrender/sw_compositor"]
