+++ rustfmt.toml
@@ -27,8 +27,6 @@ merge_derives = true
 newline_style = "Unix"
 normalize_comments = false
 reorder_imports = true
-report_fixme = "Never"
-report_todo = "Never"
 skip_children = false
 space_after_colon = true
 space_before_colon = false
